src/requirements.txt
Django==2.0
django-crispy-forms==1.7.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.1
pytz==2018.4
